Achivements data
This file is generated by MyData. The Neil exporter has not yet been put to the test to the Neil compiler, so errors cannot yet be sure not to be present. Except for fixes if needed no drastic changes should be expected as this all uses the same data as the original game.