Byebye Game Jolt - closes #51
This has not yet been fully tested, but I don't expect much trouble. The data gathered was never used in the game anywhere as the decission to cut Game Jolt from the game came pretty soon, even before the basis was made for the achievement system.

The old game still supports the Game Jolt API, but not for long anymore. An exact date for discontinuation has not yet been set, but if things go according to planning it's likely gonna happen before fall... Likely even far before that.