Bundle switch
Map has also been set to Scyndi.
Most if it is now dummied. This is because there is too much to set up here that is incompatible with SCI.
SCI also has much more efficient ways to handle Kthura maps, meaning that nearly all functions here will have to be rewritten anyway. #171